# Speed up testing against IUSE by not using regexp